Struct tl_lang_syn::cursor::Cursor

pub struct Cursor<'a> { /* fields omitted */ }

A cheaply copyable cursor into a &str.

This cursor holds a shared reference into the immutable str and can be efficiently manipulated and copied around.

A Cursor is created from an input &str. Upon creation it will point to the first byte of the string.

Two cursors are equal if they have the same location in the same input string, and have the same scope.

Methods

impl<'a> Cursor<'a>

pub fn new(input: &'a str) -> Cursor<'a>

Create a new Cursor from the given input. The created cursor points to the first byte of the string.

pub fn span(self) -> Span

Return the span of the whole cursor.

pub fn offset(self) -> usize

Get the offset this Cursor is located at.

pub fn to_str(self) -> &'a str

Get the part of the input string this cursor is operating on.