Trait tl_lang_syn::print::Print

pub trait Print: Sealed {
    fn print(&self, f: &mut Formatter) -> Result;

    fn display_wrapper<'a>(&'a self) -> DisplayWrapper<'a, Self> { ... }
}

Common printing facility for syntax tree nodes.

This trait is sealed and cannot be implemented for types outside of tl-lang-syn to avoid breaking backwards compatibility when adding new methods or derived traits.

Required Methods

fn print(&self, f: &mut Formatter) -> Result

Provided Methods

fn display_wrapper<'a>(&'a self) -> DisplayWrapper<'a, Self>