Struct tl_lang_syn::punctuated::Punctuated

pub struct Punctuated<T, P> { /* fields omitted */ }

A punctuated sequence of syntax tree nodes of type T separated by punctuation of type P.

Methods

impl<T, P> Punctuated<T, P> where
T: Synom,
P: Synom,

pub fn parse<'a>(
input: Cursor<'a>,
trailing_punct: TrailingPunctuation,
count: Count,
whitespace: Whitespace
) -> IResult<Cursor<'a>, Self>

Parse syntax tree nodes with punctuation in between using 3 configuration options:

TrailingPunctuation — whether parse must try to consume a trailing punctuation if any.

Count — how many syntax tree nodes parse is allowed to consume.

Whitespace — whether parse should ignore whitespace.

impl<T, P> Punctuated<T, P>

pub fn new() -> Punctuated<T, P>

Create an empty punctuated sequence.

pub fn is_empty(&self) -> bool

Determine whether this punctuated sequence is empty, meaning it contains no syntax tree nodes or punctuation.

pub fn len(&self) -> usize

Return the number of syntax tree nodes in this punctuated sequence.

This is the number of nodes of type T, not counting the punctuation of type P.

pub fn first(&self) -> Option<Pair<&T, &P>>

Borrow the first punctuated pair in this sequence.

pub fn first_mut(&mut self) -> Option<Pair<&mut T, &mut P>>

Mutably borrow the first punctuated pair in this sequence.

pub fn last(&self) -> Option<Pair<&T, &P>>

Borrow the last punctuated pair in this sequence.

pub fn last_mut(&mut self) -> Option<Pair<&mut T, &mut P>>

Mutably borrow the last punctuated pair in this sequence.

Important traits for Iter<'a, T, P>
impl<'a, T, P> Iterator for Iter<'a, T, P> type Item = &'a T;

pub fn iter(&self) -> Iter<T, P>

Return an iterator over borrowed syntax tree nodes of type &T.

Important traits for IterMut<'a, T, P>
impl<'a, T, P> Iterator for IterMut<'a, T, P> type Item = &'a mut T;

pub fn iter_mut(&mut self) -> IterMut<T, P>

Return an iterator over mutably borrowed syntax tree nodes of type &mut T.

Important traits for Pairs<'a, T, P>
impl<'a, T, P> Iterator for Pairs<'a, T, P> type Item = Pair<&'a T, &'a P>;

pub fn pairs(&self) -> Pairs<T, P>

Return an iterator over the contents of this sequence as borrowed punctuated pairs.

Important traits for PairsMut<'a, T, P>
impl<'a, T, P> Iterator for PairsMut<'a, T, P> type Item = Pair<&'a mut T, &'a mut P>;

pub fn pairs_mut(&mut self) -> PairsMut<T, P>

Return an iterator over the contents of this sequence as mutably borrowed punctuated pairs.

Important traits for IntoPairs<T, P>
impl<T, P> Iterator for IntoPairs<T, P> type Item = Pair<T, P>;

pub fn into_pairs(self) -> IntoPairs<T, P>

Return an iterator over the contents of this sequence as owned punctuated pairs.

pub fn push_value(&mut self, value: T)

Append a syntax tree node onto the end of this punctuated sequence. The sequence must previously have a trailing punctuation.

Use push instead if the punctuated sequence may or may not already have trailing punctuation.

Panics

This method panics if the sequence does not already have a trailing punctuation when this method is called.

pub fn push_punct(&mut self, punctuation: P)

Append a trailing punctuation onto the end of this punctuated sequence. The sequence must be non-empty and must not already have trailing punctuation.

Panics

This method panics if the sequence is empty or already has a trailing punctuation.

pub fn pop(&mut self) -> Option<Pair<T, P>>

Remove the last punctuated pair from this sequence, or None if the sequence is empty.

pub fn remove(&mut self, index: usize) -> Pair<T, P>

Remove an element at position index.

Panics

This method panics if index is greater than or equal to the number of elements previously in this punctuated sequence.

pub fn empty_or_trailing(&self) -> bool

Return true if either this Punctuated is empty, or it has a trailing punctuation.

impl<T, P> Punctuated<T, P> where
P: Default,

pub fn push(&mut self, value: T)

Append a syntax tree node onto the end of this punctuated sequence.

If there is not a trailing punctuation in this sequence when this method is called, the default value of punctuation type P is inserted before the given value of type T.

pub fn insert(&mut self, index: usize, value: T)

Insert an element at position index.

Panics

This method panics if index is greater than the number of elements previously in this punctuated sequence.