Trait tl_lang_syn::spanned::Spanned

pub trait Spanned: Sealed {
    fn span(&self) -> Span;
}

A trait that can provide the Span of the complete contents of a syntax tree node.

This trait is sealed and cannot be implemented for types outside of tl-lang-syn to avoid breaking backwards compatibility when adding new methods or derived traits.

Required Methods

fn span(&self) -> Span

Return a Span covering the complete contents of this syntax tree node, or Span::zeroed() if this node is empty.