Trait tl_lang_syn::synom::Synom

pub trait Synom: Sized + Sealed {
    fn parse_cursor<'a>(input: Cursor<'a>) -> IResult<Cursor<'a>, Self>;

    fn parse_str(input: &str) -> IResult<&str, Self> { ... }
}

Parsing interface implemented by all types that can be parsed in a default way from a string.

This trait is sealed and cannot be implemented for types outside of tl-lang-syn to avoid breaking backwards compatibility when adding new methods or derived traits.

Required Methods

fn parse_cursor<'a>(input: Cursor<'a>) -> IResult<Cursor<'a>, Self>

Provided Methods

fn parse_str(input: &str) -> IResult<&str, Self>